Add ADKServer and AgentManager classes for agent management
